requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.3
python-dotenv==1.0.1
Flask-Cors==4.0.1
httpx==0.25.0
python-telegram-bot==20.6
pandas==2.2.3
reportlab==4.2.2
Pillow==11.0.0
openpyxl==3.1.5
requests==2.32.3
